% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fit3Trees.R
\name{fit3Trees}
\alias{fit3Trees}
\title{Fit random intercept and slope mixed effects 3Trees models}
\usage{
fit3Trees(
  Y,
  X,
  gr,
  covLin = NULL,
  covT1 = NULL,
  covT2 = NULL,
  covT3 = NULL,
  niter = 50,
  re_form = "1|gr",
  selective.inference = FALSE,
  cp = c(1e-04, 1e-04, 1e-04),
  maxd = c(3, 3, 3),
  nmin = 10,
  prec = 1e-04,
  extended = FALSE,
  seed = 111
)
}
\arguments{
\item{Y}{A vector of the outcome variable (continuous).}

\item{X}{A data frame or matrix containing the all the predictors, with labelled columns.}

\item{gr}{A vector or matrix representing the grouping variable(s). If selective.inference=TRUE, gr must be a numeric vector.}

\item{covLin}{A vector of predictor names to be included in the linear part. Default is NULL.}

\item{covT1}{A vector of predictor names to be included in the first tree. Default is NULL.}

\item{covT2}{A vector of predictor names to be included in the second tree. Default is NULL.}

\item{covT3}{A vector of predictor names to be included in the third tree. Default is NULL.}

\item{niter}{The maximum number of iterations.}

\item{re_form}{The formula for the random effect, as in \code{\link{lme4::lmer}}. Default is "(1|gr)".}

\item{selective.inference}{TRUE if post-selection inference is required; default is FALSE.}

\item{cp}{A vector of length 3 containing the rpart complexity parameters cp for each tree.}

\item{maxd}{A vector of length 3 containing the rpart maxdepth parameter for each tree.}

\item{nmin}{The minbucket setting for \code{\link{rpart::rpart}}. Assumed equal for all the trees.}

\item{prec}{The tolerance for the iterative tree search.}

\item{extended}{When TRUE, it provides the final model matrix as output.}

\item{seed}{The seed to set for selective inference data splitting.}
}
\value{
A fitted mixed effects model object, containing the following elements:
\item{mse.final}{The mean squared error of the final model.}
\item{mse.random}{The mean squared error of the model with random effects.}
\item{mod.final}{The fitted mixed effects model object.}
\item{Tree1}{The \code{\link{rpart::rpart}} tree object for Tree1.}
\item{Tree2}{The \code{\link{rpart::rpart}} tree object for Tree2.}
\item{Tree3}{The \code{\link{rpart::rpart}} tree object for Tree3.}
\item{modmat}{The model matrix with tree predictions (if extended is TRUE).}
}
\description{
It can use a predictive or a post selection inference approach
}
\examples{

data(mydat)
X<-mydat[,1:5]
mod3T <- fit3Trees(Y=mydat$Y, X=X, gr=mydat$gr, covLin=names(X), 
                 covT1=names(X[,1:3]), covT2=names(X[,4:5]), 
                 covT3=colnames(X),
                 niter = 50, re_form = "(1|gr)",
                 selective.inference = FALSE)
}
\seealso{
\code{\link{lme4::lmer}}, \code{\link{rpart::rpart}}
}
