% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MSE3T.R
\name{MSE3T}
\alias{MSE3T}
\title{Calculate MSE, RMSE, and MAE for 3Trees models.}
\usage{
MSE3T(mod3T, newdata, re = TRUE, save = FALSE)
}
\arguments{
\item{mod3T}{The output of fit3trees.}

\item{newdata}{New data to evaluate the predictions.}

\item{re}{Logical indicating whether to include random effects in the prediction. Default is TRUE.}

\item{save}{Logical indicating whether to return the metrics as a list. Default is FALSE.}
}
\value{
If save is TRUE, a list of MSE, RMSE, and MAE. Otherwise, the metrics are printed.
}
\description{
Calculate MSE, RMSE, and MAE for 3Trees models.
}
