% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perm3T.R
\name{perm3T}
\alias{perm3T}
\title{Evaluate random intercept mixed effects 3Trees models when permuting the order of the three trees.}
\usage{
perm3T(
  Y,
  X,
  gr,
  covLin = NULL,
  covT1 = NULL,
  covT2 = NULL,
  covT3 = NULL,
  niter = 50,
  re_form = "(1|gr)",
  cp = 1e-04,
  maxd = 3,
  nmin = 10,
  prec = 1e-04,
  seed = 111,
  trainprop = 2/3
)
}
\arguments{
\item{Y}{A vector of the outcome variable (continuous).}

\item{X}{A data frame or matrix containing the all the predictors, with labelled columns.}

\item{gr}{A vector or matrix representing the grouping variable(s). If selective.inference=TRUE, gr must be a numeric vector.}

\item{covLin}{A vector of predictor names to be included in the linear part. Default is NULL.}

\item{covT1}{A vector of predictor names to be included in the first tree. Default is NULL.}

\item{covT2}{A vector of predictor names to be included in the second tree. Default is NULL.}

\item{covT3}{A vector of predictor names to be included in the third tree. Default is NULL.}

\item{niter}{The maximum number of iterations.}

\item{re_form}{The formula for the random effect, as in lme4. Default is "(1|gr)".}

\item{cp}{The rpart complexity parameters cp, assumed equal all trees.}

\item{maxd}{The rpart maxdepth parameter, assumed equal all trees.}

\item{nmin}{The minbucket setting for rpart, assumed equal all trees.}

\item{prec}{The tolerance for the iterative tree search.}

\item{seed}{The seed to set for train-and-test data splitting.}

\item{trainprop}{Proportion of units in the train set.}
}
\value{
A matrix containing the MSE values for different combinations of permuted trees.
}
\description{
It splits the data into train (trainprop  * n) and test (the rest).
}
\examples{
data(mydat)
X<-mydat[,1:5]
mse_matrix <- perm3T(Y=mydat$Y, X, gr=mydat$gr, covLin = NULL, covT1 = names(X[,1:3]), 
                     covT2 = names(X[,4:5]), covT3 = colnames(X),
                     niter = 50, re_form = "(1|gr)", cp = 0.0001, maxd = 3, nmin = 10,
                     prec = 1e-4, seed = 111, trainprop = 2/3)
 mse_matrix                    

}
