% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot3T.R
\name{plot3T}
\alias{plot3T}
\title{Plot a tree of mixed effects 3Trees models.}
\usage{
plot3T(mod3T, tree = 1)
}
\arguments{
\item{mod3T}{The output of fit3trees.}

\item{tree}{Tree indicator: Must be 1, 2, or 3.}
}
\value{
A plot for one of the tree components
}
\description{
Plot a tree of mixed effects 3Trees models.
}
\examples{
data(mydat)
X<-mydat[,1:5]
mod3T <- fit3Trees(Y=mydat$Y, X=X, gr=mydat$gr, covLin=names(X), 
                 covT1=names(X[,1:3]), covT2=names(X[,4:5]), 
                 covT3=colnames(X),
                 niter = 50, re_form = "(1|gr)",
                 selective.inference = FALSE)
                 
plot3T(mod3T, tree = 1)
plot3T(mod3T, tree = 2)
plot3T(mod3T, tree = 3)

}
