% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predictT3.R
\name{predictT3}
\alias{predictT3}
\title{Predicts the outcome using random intercept and slope mixed effects 3Trees models.}
\usage{
predictT3(mod3T, newdata, re = TRUE)
}
\arguments{
\item{mod3T}{The output of fit3trees.}

\item{newdata}{New data to which to apply the prediction.}

\item{re}{Logical indicating whether to include random effects in the prediction. Default is TRUE.}
}
\value{
Predicted outcomes based on the fitted model.
}
\description{
Predicts the outcome using random intercept and slope mixed effects 3Trees models.
}
