% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prune3T.R
\name{prune3T}
\alias{prune3T}
\title{Prune a random intercept and slope mixed effects 3Trees model}
\usage{
prune3T(
  mod3T,
  cutT1 = NULL,
  cutT2 = NULL,
  cutT3 = NULL,
  cutLin = NULL,
  re_form = "1|gr"
)
}
\arguments{
\item{mod3T}{The output of fit3trees estimated setting extended=TRUE}

\item{cutT1}{A list of final nodes from Tree 1 to merge for pruning}

\item{cutT2}{A list of final nodes from Tree 2 to merge for pruning}

\item{cutT3}{A list of final nodes from Tree 3 to merge for pruning}

\item{cutLin}{A list of variables from covLin to omit}

\item{re_form}{The formula for the random effect, as in lme4, to be used to fit the pruned model. Default is "(1|gr)".}
}
\value{
A fitted mixed effects model object.
}
\description{
Prune a random intercept and slope mixed effects 3Trees model
}
\examples{
data(mydat)
X<-mydat[,1:5]
mod3T <- fit3Trees(Y=mydat$Y, X=X, gr=mydat$gr, covLin=names(X), 
                 covT1=names(X[,1:3]), covT2=names(X[,4:5]), 
                 covT3=colnames(X),
                 niter = 50, re_form = "(1|gr)",
                 selective.inference = FALSE, extended=TRUE)
                 
mod3T.pruned <- prune3T(mod3T, cutT1=list(c(1,2)), cutT2=NULL, cutT3=NULL, cutLin=NULL)
summary(mod3T.pruned)
}
